[release/3.1] Add directory creation method that takes an ACL
