+++ src/Common/src/Microsoft/Win32/SafeHandles/SafeFindHandle.Windows.cs
